Reference

API Reference

Hydro power model

HydroPowerModels.hydrothermaloperation — Function.

hydrothermaloperation(alldata::Array{Dict{Any,Any}}, params::Dict)

Create a hydrothermal power operation model containing the policygraph the system data and the planning parameters.

Required parameters are:

data is a dict with all information of the problem.

param is a dict containing solution parameters.

source

System data and parameters

HydroPowerModels.parse_file_json — Function.

Read hydro description json file.

source

HydroPowerModels.create_param — Function.

create_param(;stages::Int = 1,
    model_constructor_grid = DCPPowerModel,
    model_constructor_grid_backward = model_constructor_grid,
    model_constructor_grid_forward = model_constructor_grid_backward,
    post_method = PowerModels.post_opf,
    optimizer = with_optimizer(GLPK.Optimizer),
    optimizer_backward = optimizer,
    optimizer_forward = optimizer_backward,
    setting = Dict("output" => Dict("branch_flows" => true,"duals" => true)),
    verbose = false,
    stage_hours = 1)

Create Parameters Dictionary.

Keywords are:

stages::Int : Number of stages.

modelconstructorgrid : Network formulation (Types from https://github.com/lanl-ansi/PowerModels.jl).

optimizer : Optimizer factory (http://www.juliaopt.org/JuMP.jl/v0.19.0/solvers/).

setting : PowerModels settings (https://github.com/lanl-ansi/PowerModels.jl/blob/e28644bf85232a5322adeeb847c0d18b7ff4f235/src/core/base.jl#L6-L34)) .

verbose : Boolean to indicate information prints.

stage_hours : Number of hours in each stage.

source

Training the policy

HydroPowerModels.train — Function.

train(hydromodel::HydroPowerModel;kwargs...)

Train future cost function using SDDP.

Keyword arguments (same as SDDP.train): https://github.com/odow/SDDP.jl/blob/0490bea2c46787e1d4d63a5491ea0106c7fe70cf/src/algorithm.jl#L780-L827

source

Visualize data and results

HydroPowerModels.plotscenarios — Function.

plotscenarios(scen::Array{Float64,2}; savepath::String ="",
    save::Bool = false, fileformat::String = "png", kwargs...)

Plots a set of scenarios.

Parameters:

scen : Scenarios matrix (Stages x Scenarious).

save : Bool to indicate if figure is to be saved.

savepath : Path save figure.

fileformat : Figure file format.

kwargs : Aditional keyword arguments for plot function.

source

HydroPowerModels.plot_grid — Function.

plot_grid(data::Dict;path=nothing,size_fig = [15cm, 15cm],node_label=false,nodelabeldist=4.5)

Plot Grid installed Power.

Paremeters:

data : HydroPowerModel single stage data.

path : Path to save grid plot.

size_fig : Size figure.

node_label : Plot nodel label on grid.

nodelabeldist: Nodel label distance from node.

source

HydroPowerModels.plot_aggregated_results — Function.

plotaggregatedresults(results::Dict)

Plot Aggregated Results. Figures are of aggregated quantities, but the methods used to aggregate were chosen in order to help analysis. For example: The final nodal price is an average of nodal prices weighted by the contribution of local loads to the total demand; Reservoir volume was grouped weighted by the amount of energy that could be produced by the stored water (as was the inflow of water).

Paremeter:

results: Simulation results.

source

HydroPowerModels.plotresults — Function.

plotresults(results::Dict;nc::Int = 3)

Common Plots.

Parameters:

results : Simulations output.

nc : Number of figures per line.

source

HydroPowerModels.plot_bound — Function.

plot_bound(m)

Plots the SDDP outer bound per iteration.

source