API Reference

Hydro power model

HydroPowerModels.hydrothermaloperation — Function

hydrothermaloperation(alldata::Array{Dict{Any,Any}}, params::Dict)

Create a hydrothermal power operation model containing the policygraph the system data and the planning parameters.

Required parameters are:

alldata is a vector of dicts with information of the problem's stages.

param is a dict containing solution parameters.

System data and parameters

HydroPowerModels.parse_file_json — Function

Read hydro description json file.

HydroPowerModels.create_param — Function

create_param(;stages::Int = 1,
    model_constructor_grid = DCPPowerModel,
    model_constructor_grid_backward = model_constructor_grid,
    model_constructor_grid_forward = model_constructor_grid_backward,
    post_method = PowerModels.build_opf,
    optimizer = GLPK.Optimizer,
    optimizer_backward = optimizer,
    optimizer_forward = optimizer_backward,
    setting = Dict("output" => Dict("branch_flows" => true,"duals" => true)),
    verbose = false,
    stage_hours = 1)

Create Parameters Dictionary.

Keywords are:

stages::Int : Number of stages.

modelconstructorgrid : Network formulation (Types from https://github.com/lanl-ansi/PowerModels.jl).

optimizer : Optimizer factory (http://www.juliaopt.org/JuMP.jl/v0.19.0/solvers/).

setting : PowerModels settings (https://github.com/lanl-ansi/PowerModels.jl/blob/e28644bf85232a5322adeeb847c0d18b7ff4f235/src/core/base.jl#L6-L34)) .

verbose : Boolean to indicate information prints.

stage_hours : Number of hours in each stage.

Training the policy

HydroPowerModels.train — Function

train(hydromodel::HydroPowerModel;kwargs...)

Train future cost function using SDDP.

Keyword arguments (same as SDDP.train): https://github.com/odow/SDDP.jl/blob/0490bea2c46787e1d4d63a5491ea0106c7fe70cf/src/algorithm.jl#L780-L827

Visualize data and results

Missing docstring.

Missing docstring for HydroPowerModels.plotscenarios. Check Documenter's build log for details.

Missing docstring.

Missing docstring for HydroPowerModels.plot_grid. Check Documenter's build log for details.

Missing docstring.

Missing docstring for HydroPowerModels.plot_aggregated_results. Check Documenter's build log for details.

Missing docstring.

Missing docstring for HydroPowerModels.plotresults. Check Documenter's build log for details.

Missing docstring.

Missing docstring for HydroPowerModels.plot_bound. Check Documenter's build log for details.